Synchronization: Fences (CPU-GPU) and Semaphores (GPU-GPU) (#53)
* added llri::flags<E>

* fixed some issues with flags<>

* fence API

* moved flags to flags.hpp

* added copyright

* added timout variables

* DX12 implementation

* fixed memory leak in dx12 implementation

* vulkan implementation and api tweaks

* helped intellisense

* implemented fence unit tests

* minor documentation update

* llri::Semaphore

* indenting fix

* renamed U to underlying_t as per https://github.com/Rythe-Interactive/Legion-LLRI/pull/53#discussion_r711776183

* switched to WaitForMultipleObjects for DX12 fences

* renamed timeout definitions

+++ legion/engine/llri/detail/command_group.inl
@@ -16,7 +16,7 @@ namespace LLRI_NAMESPACE
         {
             if (cmdList->queryState() == command_list_state::Recording)
             {
-                const std::string str = "CommandGroup::reset() returned ErrorInvalidState because CommandList " + std::to_string((int)cmdList) + " was in the command_list_state::Recording state.";
+                const std::string str = "CommandGroup::reset() returned ErrorInvalidState because CommandList " + std::to_string((uint64_t)cmdList) + " was in the command_list_state::Recording state.";
                 m_validationCallback(validation_callback_severity::Error, validation_callback_source::Validation, str.c_str());
                 return result::ErrorInvalidState;
             }
